[JE2-252] Rename PieTask
